+++ src/index.ts
@@ -1,5 +1,4 @@
 import {
-  ChatCompletionRequestMessageRoleEnum,
   ChatCompletionResponseMessageRoleEnum,
   Configuration,
   OpenAIApi
@@ -23,6 +22,7 @@ import { Message } from './messages/Message'
 import { mongoClient } from './mongo/MongoClient'
 import { Ntfy } from './ntfy/Ntfy'
 import { log } from './logger'
+import { queueMessage } from './messages/index'
 
 const API_KEY = getEnv('API_KEY')
 const DISCORD_TOKEN = getEnv('DISCORD_TOKEN')
@@ -31,14 +31,14 @@ const NTFY_TOPIC = getEnv('NTFY_TOPIC')
 
 export const ntfy = new Ntfy(NTFY_TOPIC)
 
-const openAiHelper = new OpenAiHelper(
+export const openAiHelper = new OpenAiHelper(
   new OpenAIApi(
     new Configuration({
       apiKey: API_KEY
     })
   )
 )
-const client = new Client({
+export const client = new Client({
   intents: [
     GatewayIntentBits.Guilds,
     GatewayIntentBits.GuildMessages,
@@ -46,8 +46,7 @@ const client = new Client({
   ]
 })
 
-// const channels = new Map<string, Channel>()
-const guilds = new Map<string, Guild>();
+export const guilds = new Map<string, Guild>();
 
 // eslint-disable-next-line @typescript-eslint/no-floating-promises
 (async () => {
@@ -277,7 +276,7 @@ commands.set('who', async (interaction) => {
     `\`IGNORE_EVERYONE_MENTIONS\`: ${channel.config.IGNORE_EVERYONE_MENTIONS}\n` +
     `\`LANGUAGE_MODEL\`: ${channel.config.LANGUAGE_MODEL}` +
     `\n\nSystem messages (${messages.length}): \n • ${messageList}`
-  
+
   await interaction.reply({
     content: response.slice(0, 2000),
     ephemeral: true
@@ -505,7 +504,7 @@ commands.set('system', async (interaction) => {
 
     const response = `System messages (${messages.length}): \n • ${messageList}`
     await interaction.reply({
-      content: response.slice(0, 2000),
+      content: response.slice(0, 2000)
     })
     return
   }
@@ -593,304 +592,7 @@ client.on(Events.InteractionCreate, async (interaction) => {
   }
 })
 
-client.on(Events.MessageCreate, async (message) => {
-  try {
-    if (message.guildId == null) return
-    if (client.user?.id == null || message.channelId == null) return
-    if (message.author.id === client.user.id) return
-
-    log({
-      guildId: message.guildId,
-      channelId: message.channelId,
-      message: `${message.author.username}: ${message.content}`
-    })
-
-    let guild = guilds.get(message.guildId)
-    if (guild == null) {
-      log({
-        guildId: message.guildId,
-        channelId: message.channelId,
-        message: 'Guild not found, creating new one'
-      })
-      guild = new Guild(message.guildId, DEFAULT_GUILD_CONFIG)
-      guilds.set(message.guildId, guild)
-      await guild.save()
-      void ntfy.publish('New Server', 'A new server has been added to the bot', 'partying_face')
-    }
-
-    const channel = await guild.getChannel(message.channelId)
-
-    if (channel == null) {
-      if (message.mentions.users.has(client.user.id)) {
-        log({
-          guildId: message.guildId,
-          channelId: message.channelId,
-          message: 'Channel not found, but mentioned, replying with /config message'
-        })
-        await message.reply(
-          'This channel is not configured for chatting with me.  ' +
-        'Someone with permissions needs to run `/config` to configure this channel.')
-      }
-      return
-    }
-
-    const newMessage = new Message({
-      id: message.id,
-      guildId: message.guildId,
-      channelId: message.channelId,
-      userId: message.author.id,
-      timestamp: message.createdTimestamp,
-      type: ChatCompletionRequestMessageRoleEnum.User,
-      content: message.content,
-      chatCompletionRequestMessage: {
-        role: ChatCompletionRequestMessageRoleEnum.User,
-        content: message.content,
-        name: message.author.id.toString()
-      }
-    })
-
-    await channel.addMessage(newMessage)
-
-    // also ignore @everyone or role mentions
-    if ((message.mentions.everyone || message.mentions.roles.size > 0) &&
-      channel.config.IGNORE_EVERYONE_MENTIONS) return
-
-    // also ignore all bots.. we don't want to get into a loop
-    if (message.author.bot && channel.config.IGNORE_BOTS) return
-    if (!message.mentions.has(client.user) && channel.config.ONLY_RESPOND_TO_MENTIONS) {
-      return
-    }
-
-    // if there were mentions that don't include us, ignore
-    if (message.mentions.users.size > 0 && !message.mentions.has(client.user)) {
-      log({
-        guildId: message.guildId,
-        channelId: message.channelId,
-        message: 'Message has mentions, but not us, ignoring'
-      })
-      return
-    }
-
-    await message.channel.sendTyping()
-    // continulously send typing while waiting for the completion
-    const typingInterval = setInterval(() => {
-      message.channel
-        .sendTyping()
-        .then(() => {
-          log({
-            guildId: message.guildId ?? undefined,
-            channelId: message.channelId,
-            message: 'Sent typing...'
-          })
-        })
-        .catch(() => {
-          log({
-            guildId: message.guildId ?? undefined,
-            channelId: message.channelId,
-            message: 'Error sending typing',
-            level: 'error'
-          })
-        })
-    }, 5000)
-
-    try {
-      // ensure the message is appropriate
-      const badMessages = await openAiHelper.findFlaggedMessages(
-        channel.messages.map((message) => message.content)
-      )
-
-      if (badMessages.length > 0) {
-        // remove any messages that were flagged
-        badMessages.forEach((badMessage) => {
-          const message = channel.messages.find((message) => message.content === badMessage)
-          if (message != null) {
-            log({
-              guildId: message.guildId,
-              channelId: message.channelId,
-              message: `Message flagged as inappropriate: ${message.content}`
-            })
-            void message.delete()
-          }
-        })
-
-        channel.messages = channel.messages.filter(
-          (message) => !badMessages.includes(message.content)
-        )
-
-        clearInterval(typingInterval)
-        await message.reply(channel.config.MODERATION_VIOLATION)
-        log({
-          guildId: message.guildId,
-          channelId: message.channelId,
-          message: 'Moderation violation! Ignoring message'
-        })
-        return
-      }
-
-      // lastly ensure the guild is at least one week old to prevent abuse
-      if (((message.guild?.createdTimestamp) == null) ||
-        message.guild.createdTimestamp > Date.now() - 1000 * 60 * 60 * 24 * 7) {
-        log({
-          guildId: message.guildId,
-          channelId: message.channelId,
-          message: 'Server is less than one week old, ignoring message'
-        })
-        await message.reply('Sorry. To prevent abuse, your server must be at ' +
-          'least one week old to use this bot')
-        clearInterval(typingInterval)
-        return
-      }
-
-      if (channel.config.LANGUAGE_MODEL.toLowerCase() === 'gpt-4') {
-        if (guild.gpt4TokensAvailable <= 0) {
-          log({
-            guildId: message.guildId,
-            channelId: message.channelId,
-            message: 'No tokens available for GPT-4'
-          })
-          await message.reply('Sorry, you have run out of GPT-4 tokens. `/tokens` to get more or ' +
-        '`/config` to switch to GPT-3')
-          clearInterval(typingInterval)
-          guild.gpt4TokensAvailable = 0
-          await guild.save()
-          return
-        }
-      } else {
-        if (guild.gpt3TokensAvailable <= 0) {
-          log({
-            guildId: message.guildId,
-            channelId: message.channelId,
-            message: 'No tokens available for GPT-3'
-          })
-          await message.reply('Sorry, you have run out of GPT-3 tokens. `/tokens` to get more')
-          guild.gpt3TokensAvailable = 0
-          clearInterval(typingInterval)
-          await guild.save()
-          return
-        }
-      }
-
-      log({
-        guildId: message.guildId,
-        channelId: message.channelId,
-        message: `Creating completion with prompt tokens: ${channel.countTotalTokens()}`
-      })
-
-      let response = await openAiHelper.createChatCompletion(
-        channel.messages.map((message) => message.chatCompletionRequestMessage),
-        channel.config.LANGUAGE_MODEL,
-        message.author.id
-      )
-
-      // subtract from the guild's token count
-      if (channel.config.LANGUAGE_MODEL.toLowerCase() === 'gpt-4') {
-        try {
-          await guild.subtractGpt4Tokens(channel.countTotalTokens())
-        } catch (error) {
-          log({
-            guildId: message.guildId,
-            channelId: message.channelId,
-            message: 'Error subtracting tokens' + (error as string),
-            level: 'error'
-          })
-          return
-        }
-      } else {
-        try {
-          await guild.subtractGpt3Tokens(channel.countTotalTokens())
-        } catch (error) {
-          log({
-            guildId: message.guildId,
-            channelId: message.channelId,
-            message: 'Error subtracting tokens' + (error as string),
-            level: 'error'
-          })
-          return
-        }
-      }
-
-      // let's ensure our own response doesn't violate any moderation
-      // rules
-      if ((await openAiHelper.findFlaggedMessages([response])).length > 0) {
-        log({
-          guildId: message.guildId,
-          channelId: message.channelId,
-          message: 'Moderation violation (from our own response...): ' + response
-        })
-        clearInterval(typingInterval)
-        await message.reply(channel.config.MODERATION_VIOLATION)
-        return
-      }
-
-      const newMessage = new Message({
-        guildId: message.guildId,
-        channelId: message.channelId,
-        userId: client.user.id,
-        timestamp: Date.now(),
-        type: ChatCompletionRequestMessageRoleEnum.Assistant,
-        content: response,
-        chatCompletionRequestMessage: {
-          role: ChatCompletionRequestMessageRoleEnum.Assistant,
-          content: response,
-          name: client.user.id.toString()
-        }
-      })
-
-      await channel.addMessage(newMessage)
-
-      log({
-        guildId: message.guildId,
-        channelId: message.channelId,
-        message: 'AI Response: ' + response
-      })
-
-      // if this is the first assistant message, send the disclaimer first
-      if (!channel.disclaimerSent && channel.config.DISCLAIMER.length > 0) {
-        channel.setDisclaimerSent(true)
-        response = channel.config.DISCLAIMER + '\n\n' + response
-        log({
-          guildId: message.guildId,
-          channelId: message.channelId,
-          message: 'Adding Disclaimer to message'
-        })
-        await channel.save()
-      }
-
-      // if the message is too long, split it up into max of 2000
-      // characters per message
-      // split up the messsage into each 2000 character chunks
-      const chunks = response.match(/[\s\S]{1,2000}/g)
-      clearInterval(typingInterval)
-      for (const chunk of chunks ?? []) {
-      // if it was a mention, reply to the message
-        if (message.mentions.has(client.user)) {
-          await message.reply(chunk)
-        } else {
-        // otherwise, send message to channel
-        // (should only happen if ONLY_RESPOND_TO_MENTIONS = FALSE)
-          await message.channel.send(chunk)
-        }
-      }
-    } catch (e) {
-      log({
-        guildId: message.guildId,
-        channelId: message.channelId,
-        message: 'Error: ' + (e as string),
-        level: 'error'
-      })
-      await message.reply(channel.config.ERROR_RESPONSE)
-    } finally {
-      clearInterval(typingInterval)
-    }
-  } catch (e) {
-    log({
-      guildId: message.guildId ?? undefined,
-      channelId: message.channelId,
-      message: 'Error: ' + (e as string),
-      level: 'error'
-    })
-  }
-})
+client.on(Events.MessageCreate, queueMessage)
 
 // every 10 minutes, set our status to online
 setInterval(() => {